requirements.txt
annotated-types==0.6.0
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
colorama==0.4.6
distro==1.9.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
openai==1.16.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2022.7
tzlocal==4.2
